Connect the BooksList component to the store using mapStateToProps and add PropTypes
