+++ src/reducers/books.js
@@ -1,5 +1,24 @@
-import initialState from '../index';
 import { CREATE_BOOK, REMOVE_BOOK } from '../actions/index';
+
+const initialState = {
+  books: [
+    {
+      id: Math.floor(Math.random() * 100),
+      title: 'Dracula',
+      category: 'Action',
+    },
+    {
+      id: Math.floor(Math.random() * 100),
+      title: 'Romeo and Juliet',
+      category: 'Learning',
+    },
+    {
+      id: Math.floor(Math.random() * 100),
+      title: 'The Last Samurai',
+      category: 'History',
+    },
+  ],
+};
 
 const booksReducer = (state = initialState, action) => {
   switch (action.type) {
