+++ src/reducers/index.js
@@ -1,3 +1,5 @@
+/* eslint import/no-cycle: [2, { ignoreExternal: true }] */
+
 import { combineReducers, createStore } from 'redux';
 import booksReducer from './books';
